Amz_Product_Rec_Project/requirements.txt
pandas
numpy
nltk
scikit-learn
<<<<<<< HEAD
punkt
=======
streamlit
Pillow
>>>>>>> refs/remotes/origin/master
